Clean up obsolete classes
